refactor(types): Move all runtime values to `runtime-values.ts`
