+++ packages/types/src/web3.ts
@@ -12,52 +12,3 @@ export type OKXWalletWeb3Provider = 'okx_wallet';
 export type BaseWeb3Provider = 'base';
 
 export type Web3Provider = MetamaskWeb3Provider | BaseWeb3Provider | CoinbaseWalletWeb3Provider | OKXWalletWeb3Provider;
-
-/**
- * @deprecated Use `import { WEB3_PROVIDERS } from "@clerk/shared/web3"` instead.
- *
- * @hidden
- */
-export const WEB3_PROVIDERS: Web3ProviderData[] = [
-  {
-    provider: 'metamask',
-    strategy: 'web3_metamask_signature',
-    name: 'MetaMask',
-  },
-  {
-    provider: 'base',
-    strategy: 'web3_base_signature',
-    name: 'Base',
-  },
-  {
-    provider: 'coinbase_wallet',
-    strategy: 'web3_coinbase_wallet_signature',
-    name: 'Coinbase Wallet',
-  },
-  {
-    provider: 'okx_wallet',
-    strategy: 'web3_okx_wallet_signature',
-    name: 'OKX Wallet',
-  },
-];
-
-interface getWeb3ProviderDataProps {
-  provider?: Web3Provider;
-  strategy?: Web3Strategy;
-}
-
-/**
- * @deprecated This utility will be dropped in the next major release.
- *
- * @hidden
- */
-export function getWeb3ProviderData({
-  provider,
-  strategy,
-}: getWeb3ProviderDataProps): Web3ProviderData | undefined | null {
-  if (provider) {
-    return WEB3_PROVIDERS.find(p => p.provider == provider);
-  }
-
-  return WEB3_PROVIDERS.find(p => p.strategy == strategy);
-}
